Set original SQL back one version

+++ Common/src/main/resources/db/sqlite.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS "avpn_6_ip" (
 	"ip"	TEXT NOT NULL UNIQUE,
 	"type"	INTEGER NOT NULL,
 	"cascade"	INTEGER,
-	"consensus"	REAL,
+	"consensus"	INTEGER,
 	PRIMARY KEY("id" AUTOINCREMENT)
 );
